scheduler: refactor: split job_handlers by concern with active-job tracker helpers
